WIP: Patches per facet are not good

+++ Polyhedron/demo/Polyhedron/resources/shader_c3t3.vert
@@ -3,7 +3,7 @@ in vec4 vertex;
 in vec3 normals;
 in vec3 colors;
 in vec3 center;
-in uint subdomain_in;
+in float subdomain_in;
 uniform  mat4 mvp_matrix;
 uniform  mat4 mv_matrix;
 uniform mat4 norm_matrix;
@@ -12,7 +12,7 @@ uniform  float shrink_factor;
 out vec4 fP;
 out vec3 fN;
 out vec4 color;
-flat out uint subdomain_out;
+flat out float subdomain_out;
 uniform  float point_size;
 void main(void)
 {
